refactor(services/webdav): split webdav service into separate crate
