+++ core/services/webdav/src/config.rs
@@ -52,10 +52,10 @@ impl Debug for WebdavConfig {
     }
 }
 
-impl crate::Configurator for WebdavConfig {
+impl opendal_core::Configurator for WebdavConfig {
     type Builder = WebdavBuilder;
 
-    fn from_uri(uri: &crate::types::OperatorUri) -> crate::Result<Self> {
+    fn from_uri(uri: &opendal_core::OperatorUri) -> opendal_core::Result<Self> {
         let mut map = uri.options().clone();
         if let Some(authority) = uri.authority() {
             map.insert("endpoint".to_string(), format!("https://{authority}"));
@@ -76,8 +76,8 @@ impl crate::Configurator for WebdavConfig {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::Configurator;
-    use crate::types::OperatorUri;
+    use opendal_core::Configurator;
+    use opendal_core::OperatorUri;
 
     #[test]
     fn from_uri_sets_endpoint_and_root() {
