+++ core/services/webdav/src/deleter.rs
@@ -15,23 +15,34 @@
 // specific language governing permissions and limitations
 // under the License.
 
-/// Default scheme for webdav service.
-pub const WEBDAV_SCHEME: &str = "webdav";
+use std::sync::Arc;
 
-use crate::types::DEFAULT_OPERATOR_REGISTRY;
+use http::StatusCode;
 
-mod backend;
-mod config;
-mod core;
-mod deleter;
-mod error;
-mod lister;
-mod writer;
+use super::core::*;
+use super::error::parse_error;
+use opendal_core::raw::oio;
+use opendal_core::raw::*;
+use opendal_core::*;
 
-pub use backend::WebdavBuilder as Webdav;
-pub use config::WebdavConfig;
+pub struct WebdavDeleter {
+    core: Arc<WebdavCore>,
+}
 
-#[ctor::ctor]
-fn register_webdav_service() {
-    DEFAULT_OPERATOR_REGISTRY.register::<Webdav>(WEBDAV_SCHEME);
+impl WebdavDeleter {
+    pub fn new(core: Arc<WebdavCore>) -> Self {
+        Self { core }
+    }
+}
+
+impl oio::OneShotDelete for WebdavDeleter {
+    async fn delete_once(&self, path: String, _: OpDelete) -> Result<()> {
+        let resp = self.core.webdav_delete(&path).await?;
+
+        let status = resp.status();
+        match status {
+            StatusCode::NO_CONTENT | StatusCode::NOT_FOUND => Ok(()),
+            _ => Err(parse_error(resp)),
+        }
+    }
 }
