+++ core/services/webdav/src/error.rs
@@ -18,8 +18,8 @@
 use http::Response;
 use http::StatusCode;
 
-use crate::raw::*;
-use crate::*;
+use opendal_core::raw::*;
+use opendal_core::*;
 
 /// Parse error response into Error.
 pub(super) fn parse_error(resp: Response<Buffer>) -> Error {
