+++ core/services/webdav/src/lister.rs
@@ -21,8 +21,9 @@ use http::StatusCode;
 
 use super::core::*;
 use super::error::*;
-use crate::raw::*;
-use crate::*;
+use opendal_core::raw::oio;
+use opendal_core::raw::*;
+use opendal_core::*;
 
 pub struct WebdavLister {
     core: Arc<WebdavCore>,
