+++ core/services/webdav/src/writer.rs
@@ -21,8 +21,8 @@ use http::StatusCode;
 
 use super::core::*;
 use super::error::parse_error;
-use crate::raw::*;
-use crate::*;
+use opendal_core::raw::*;
+use opendal_core::*;
 
 pub struct WebdavWriter {
     core: Arc<WebdavCore>,
